requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dash==2.8.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.3
frozendict==2.3.4
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multitasking==0.0.11
numpy==1.24.2
pandas==1.5.3
plotly==5.13.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
tenacity==8.2.1
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==2.2.3
yfinance==0.2.12
zipp==3.13.0
